+++ client/src/components/common/modals/accountbook-delete-by-admin/AccountbookDeleteByAdminModal.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import GrayButton from '../buttons/GrayButton';
-import RedButton from '../buttons/RedButton';
-import ModalBackground from '../modal-background/ModalBackground';
+import GrayButton from '../../buttons/GrayButton';
+import RedButton from '../../buttons/RedButton';
+import ModalBackground from '../../modal-background/ModalBackground';
 
 interface IProps {
   cancelClick?: () => void;
